Add HA/LB endpoints for kube-apiserver
* Add auto-evaluated internal endpoints, which is either the
access_ip:443|8080 or the localhost:443|8080.
* Make the external loadbalancer_apiserver endpoint to override
the internal secure endpoint. Also clarify the loadbalancer_apiserver
vars and usecases.
* A notion for the 'kube_apiserver_multiaccess' is: ignore endpoints and
loadbalancers and use the apiservers IPs as a comma-separated
list (a default mode). Otherwise, clients shall use the given
endpoints/loadbalancers.
* Define listen mode for apiservers. When internal endpoints are provided
by a local loadbalancers, they should not listen at the localhost: 443|8080
as the loadbalancers do that.
* Define connections security mode for kube controllers, schedulers, proxies
(leave it insecure be default, which is the current deployment choice)
* Make multi-apiserver access as an optional choice
(default is false, unless it is supported by clients)
* Rework the groups['kube-master'][0] hardcode defining the apiserver
endpoints.
* Improve grouping of vars and add facts for kube_apiserver.
* Define kube_apiserver_insecure_bind_address as a fact, add more
facts for ease of use.

Signed-off-by: Bogdan Dobrelya <bdobrelia@mirantis.com>